fix rosproxy test failure (#454)
The test sometimes got stuck in a remote call `request_shutdown` which was there only to fix a situation when it is stuck. The problem was that there was a race - setting `self.quit = True` might happen while the previous request was still in `self.server.handle_request()` - when it returned, it saw that `self.quit` is `True` and quit the request handling loop. When the next request came, there was no loop to handle it, so it never returned.

The solution is to set `self.quit` in the same thread as is the checking, so there should be no race condition any more (famous last word).