subt.local_planner: rename scan_{right,left} to scan_{start,end} (#481)
The motivation for the change is to make it clearer what the arguments actually mean - "start" is the angle of the first sample in the scan and "end" is the angle of the last sample. The original names were somewhat confusing to me. The names with the default values make it clear that the default scan is 270 degrees wide and goes from right to left (from -135 to +135).