launch/x4.launch - remove lidar scan points from octomap and change filter
The motivation is that drone has to inspect the world visually in order to detect
and report artifact. LIDAR is good for obstacles detection but drone would then ignore
parts seen by LIDAR only.

The filter for removal of drone propeller is now negative in order to keep +inf range
values (our of range of RGBD camera). These "points" are then needed by octomap server.

Finally the range of sensor is now limited to 10 meters, so hopefully it will insert
free voxels without hitting the obstacles.