Step towards 3D octomap planning (#792)
* subt/octomap.py - fix crash for no frontiers situation

- reduce octomap cut to only "palette" image
- color multimap and speedup drawing
   (before 202s, after 129s for TS3 drone sample)

* pplanner.py - experimental 3D planning

* subt/octomap.py - tweak I/O to new 3d planner

- plan path 3D + 3D frontiers
- dump on verbose all map layers

- try 10 levels
- add octomap logfile drawing directly via cv2
- enable change of levels and pause replay
- draw robot position and show it in the title

* launch/x4.launch - remove lidar scan points from octomap and change filter

The motivation is that drone has to inspect the world visually in order to detect
and report artifact. LIDAR is good for obstacles detection but drone would then ignore
parts seen by LIDAR only.

The filter for removal of drone propeller is now negative in order to keep +inf range
values (our of range of RGBD camera). These "points" are then needed by octomap server.

Finally the range of sensor is now limited to 10 meters, so hopefully it will insert
free voxels without hitting the obstacles.

* add filter_pointcloud.py

* launch/x4.launch - use filter_pointclouds for filtering

* filter_pointcloud.py - first attempt to filter out propellers